requirements-simple.txt
# Simplified requirements without heavy OSM dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.31.0
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0

# For basic geometry operations (lighter than osmium)
# shapely==2.0.2  # Comment out if you get compilation errors